third_party/blink/web_tests/external/wpt/client-hints/sec-ch-width.https.html

<!DOCTYPE html>
<meta charset="utf-8">
<meta http-equiv="Delegate-CH" content="Sec-CH-Width">
<title>Tests Sec-CH-Width</title>
<link rel="help" href="https://wicg.github.io/responsive-image-client-hints/#sec-ch-width">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<script>
  setup({explicit_done: true});

  function run() {
    test(() => {
      const testImage = document.getElementsByTagName('img')[0];
      assert_equals(testImage.naturalWidth, 2 * (0.10 * window.innerWidth) * window.devicePixelRatio);
      assert_equals(testImage.naturalHeight, 3 * (0.10 * window.innerWidth) * window.devicePixelRatio);
    }, 'Sec-CH-Width should be set');
    done();
  }
</script>

<img onload="run()" src="resources/2x3-svg-scaled-by-sec-ch-width.py" sizes="10vw">
